Add cypress function `setCheckBoxValue`
https://github.com/metasfresh/metasfresh-e2e/issues/133
