+++ cypress/support/commands/byTbp.js
@@ -14,3 +14,46 @@ Cypress.Commands.add('getCurrentRecordId', () => {
     return currentRecordId;
   });
 });
+
+
+Cypress.Commands.add('setCheckBoxValue', (fieldName, isChecked, modal = false, rewriteUrl = null) => {
+  describe(`Set the Checkbox value ${fieldName} to ${isChecked}`, function () {
+
+    // the expected value is the same as the checked state
+    // (used only for verification if the checkbox has the correct value)
+    const expectedPatchValue = isChecked;
+
+    cy.getCheckboxValue(fieldName, modal).then(theCheckboxValue => {
+      if (isChecked) {
+        if (theCheckboxValue) {
+          // Nothing to do, already checked
+        } else {
+          cy.clickOnCheckBox(fieldName, expectedPatchValue, modal, rewriteUrl);
+        }
+      } else {
+        if (theCheckboxValue) {
+          cy.clickOnCheckBox(fieldName, expectedPatchValue, modal, rewriteUrl);
+        } else {
+          // Nothing to do, already unchecked
+        }
+      }
+    });
+  });
+});
+
+/**
+ * These constants should be used instead of writing strings every time
+ */
+export class RewriteURL {
+  // noinspection JSUnusedGlobalSymbols
+  /**
+   * WINDOW is the default
+   */
+  static WINDOW = '/rest/api/window/.*[^/][^N][^E][^W]$';
+
+  static PROCESS = '/rest/api/process/';
+}
+
+
+
+
